requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
python-multipart>=0.0.5
requests>=2.28.0
openai-whisper
torch
torchaudio
ffmpeg-python
aiofiles
python-slugify
jinja2